+++ src/app/loadout-builder/filter/EnergyOptions.tsx
@@ -4,7 +4,7 @@ import { t } from 'app/i18next-t';
 import { Dispatch, useCallback, useMemo } from 'react';
 import { LoadoutBuilderAction } from '../loadout-builder-reducer';
 import styles from './EnergyOptions.m.scss';
-import { loMenuSection } from './LockArmorAndPerks';
+import { loMenuSection } from './LoadoutOptimizerMenuItems';
 
 export default function EnergyOptions({
   assumeArmorMasterwork,
