+++ src/app/loadout-builder/filter/LoadoutOptimizerMenuItems.m.scss
