+++ src/app/loadout-builder/filter/LoadoutOptimizerMenuItems.m.scss.d.ts
@@ -1,0 +1,10 @@
+// This file is automatically generated.
+// Please do not change this file!
+interface CssExports {
+  'area': string;
+  'buttons': string;
+  'itemGrid': string;
+  'notItemGrid': string;
+}
+export const cssExports: CssExports;
+export default cssExports;
